requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
cftime==1.5.0
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
cycler==0.10.0
DateTime==4.3
discord-py-slash-command==3.0.1
discord.py==1.7.3
Django==3.2.6
ffmpeg==1.4
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
multidict==5.1.0
mysql==0.0.3
mysql-connector-python==8.0.26
mysqlclient==2.0.3
numpy==1.21.2
pandas==1.3.2
Pillow==8.3.1
protobuf==3.17.3
psutil==5.8.0
pycparser==2.20
pyee==8.1.0
pyerfa==2.0.0
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
python-ffmpeg==1.0.11
pytz==2021.1
requests==2.26.0
scipy==1.7.1
six==1.16.0
soupsieve==2.2.1
spotipy==2.18.0
sqlparse==0.4.1
timeago==1.0.13
tqdm==4.62.1
typing==3.7.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
xarray==0.15.1
yarl==1.6.3
youtube-dl==2021.6.6
zope.interface==5.4.0